Base admin dashboard view
Base admin dashboard view. Currently in separate file from dashboard for now. Anyone can currently access /admindb.php but to be discussed if it will be displayed in dashboard.php based on session user or in a seperate page. Also fixed up a few bugs on dashboard with current loans.